+++ include/revolution/OS/OSAlloc.h
@@ -14,7 +14,7 @@ void* OSInitAlloc(void* arenaStart, void* arenaEnd, int maxHeaps);
 OSHeapHandle OSCreateHeap(void* start, void* end);
 OSHeapHandle OSSetCurrentHeap(OSHeapHandle heap);
 
-void* OSAllocFromHeap(OSHeapHandle heap, unsigned long size);
+void* OSAllocFromHeap(OSHeapHandle heap, u32 size);
 void OSFreeToHeap(OSHeapHandle heap, void* ptr);
 
 #ifdef __cplusplus
